feat : mean mode correction to mixed unbounded-periodic kernels
Before, the mean term (k_y = 0) was identically set to zero in the mixed-unbounded kernels before taking the trasnformation. While this works in most cases where mean vorticity is 0 (like flow past a cylinder, for example), it doesnt work for the coupled fluid-solid shear-flow case in PyCFS. So we now set the mean mode to the Poisson kernel in 1D, including regularization for higher-order solvers.